refactor(api): move state and voting power computation to namespaces
